+++ redisson/src/main/java/org/redisson/RedissonSpinLock.java
@@ -199,7 +199,7 @@ public final class RedissonSpinLock extends RedissonBaseLock {
     }
 
 
-    protected RFuture<Boolean> unlockInnerAsync(long threadId, String requestId, int timeout) {
+    protected RFuture<Boolean> unlockInnerAsync(long threadId, String requestId, long timeout) {
         return evalWriteSyncedNoRetryAsync(getRawName(), LongCodec.INSTANCE, RedisCommands.EVAL_BOOLEAN,
                 "local val = redis.call('get', KEYS[2]); " +
                       "if val ~= false then " +
